{{cookiecutter.package_name}}/src/exp_000/requirements.txt
albumentations
opencv-python
numpy
torch
torchvision
torchaudio
tqdm
pandas
scikit-learn
timm
transformers
warmup_scheduler
hydra-core
wandb
matplotlib
pytorch-lightning
pytorch-gradcam
